requirements.txt
Django==3.2.13
Pillow==9.4.0

django-allauth==0.52.0
django_debug_toolbar==3.8.1
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0

flake8==6.0.0
isort==5.11.4

celery==5.2.7
redis==4.4.2
django-redis==5.2.0

stripe==5.1.0

psycopg2==2.9.5


